🐛 fix: 一覧、新規、詳細をコンポーネントに分離
