+++ src/List.js
@@ -1,19 +1,17 @@
 import { useState } from "react";
 import { useEffect } from "react";
-import Input from "./Input";
+import IndexMemo from "./IndexMemo";
+import NewMemo from "./NewMemo";
+import EditMemo from "./EditMemo";
 import "./list.css";
 
 const initialMemos = JSON.parse(localStorage.getItem("memos")) || [];
 
 export default function List() {
   const [memos, setMemos] = useState(initialMemos);
-  const [answer, setAnswer] = useState("");
   const [status, setStatus] = useState("index");
+  const [answer, setAnswer] = useState("");
   const [editingMemoId, setEditingMemoId] = useState(null);
-
-  const index = status === "index";
-  const isAdding = status === "add";
-  const isEditing = status === "edit";
 
   useEffect(() => {
     localStorage.setItem("memos", JSON.stringify(memos));
@@ -24,80 +22,48 @@ export default function List() {
     value: content.split("\n")[0],
   }));
 
-  function handleAnswerChange(e) {
-    setAnswer(e.target.value);
-  }
+  const viewForStatus = () => {
+    switch (status) {
+      case "index": {
+        return (
+          <IndexMemo
+            memos={memos}
+            firstRowOfMemos={firstRowOfMemos}
+            setStatus={setStatus}
+            setAnswer={setAnswer}
+            setEditingMemoId={setEditingMemoId}
+          />
+        );
+      }
+      case "isAdding": {
+        return (
+          <NewMemo
+            memos={memos}
+            setMemos={setMemos}
+            setStatus={setStatus}
+            answer={answer}
+            setAnswer={setAnswer}
+          />
+        );
+      }
+      case "isEditing": {
+        return (
+          <EditMemo
+            memos={memos}
+            setMemos={setMemos}
+            setStatus={setStatus}
+            editingMemoId={editingMemoId}
+            setEditingMemoId={setEditingMemoId}
+            answer={answer}
+            setAnswer={setAnswer}
+          />
+        );
+      }
+      default: {
+        // do nothing
+      }
+    }
+  };
 
-  function handleChangeStatusToAdd() {
-    setStatus("add");
-  }
-
-  async function handleAddMemo(e) {
-    e.preventDefault();
-    const nextId = memos.length > 0 ? memos[memos.length - 1].id + 1 : 0;
-    const nextMemos = [...memos, { id: nextId, content: answer }];
-    await setMemos(nextMemos);
-    await setAnswer("");
-    setStatus("index");
-  }
-
-  async function handleSaveMemo(e) {
-    e.preventDefault();
-    const targetIndex = memos.findIndex((memo) => memo.id === editingMemoId);
-    const nextMemos = [...memos];
-    nextMemos[targetIndex] = { ...nextMemos[targetIndex], content: answer };
-    await setMemos(nextMemos);
-    await setAnswer("");
-    await setEditingMemoId(null);
-    setStatus("index");
-  }
-
-  function handleMemoEditable(memo) {
-    const targetMemo = memos.find((m) => m.id === memo.id);
-    setAnswer(targetMemo.content);
-    setEditingMemoId(targetMemo.id);
-    setStatus("edit");
-  }
-
-  function handleDeleteMemo(memo) {
-    setMemos(memos.filter((memo) => memo.id !== editingMemoId));
-    setAnswer("");
-    setStatus("index");
-    setEditingMemoId(null);
-  }
-
-  if (index) {
-    return (
-      <div class="container">
-        <h1>メモ一覧</h1>
-        <ul>
-          {firstRowOfMemos.map((memo) => (
-            <li key={memo.id} onClick={() => handleMemoEditable(memo)}>
-              {memo.value}
-            </li>
-          ))}
-          <li onClick={handleChangeStatusToAdd}>+</li>
-        </ul>
-      </div>
-    );
-  } else if (isAdding) {
-    return (
-      <div class="container">
-        <h1>新規メモ</h1>
-        <Input value={answer} handleAnswerChange={handleAnswerChange} />
-        <button type="submit" onClick={handleAddMemo}>
-          add
-        </button>
-      </div>
-    );
-  } else if (isEditing) {
-    return (
-      <div class="container">
-        <h1>メモ詳細</h1>
-        <Input value={answer} handleAnswerChange={handleAnswerChange} />
-        <button onClick={handleSaveMemo}>save</button>
-        <button onClick={handleDeleteMemo}>delete</button>
-      </div>
-    );
-  }
+  return <>{viewForStatus()}</>;
 }
